fix(frontend): display error message when grid data are not loading
