+++ frontend/src/components/layout/DataTable.tsx
@@ -1,6 +1,7 @@
 import { DataGrid, GridColDef, GridToolbar } from '@mui/x-data-grid';
 import { Article } from '../../constants/types';
 import { useIsDarkMode } from '../../contexts/ThemeContext';
+import { ErrorMessage } from '../features/ErrorMessage';
 
 interface DataTableProps {
   data: Article[];
@@ -12,13 +13,9 @@ function DataTable({ data, columns, error }: Readonly<DataTableProps>) {
   const isDarkMode = useIsDarkMode();
 
   if (error) {
-    return (
-      <div className="rounded-2xl border border-red-200 bg-red-50 p-6 text-red-700 shadow-sm dark:border-red-900/60 dark:bg-red-950/40 dark:text-red-200">
-        <p className="m-0 font-semibold">Unable to load articles.</p>
-        <p className="m-0 mt-2 text-sm">{error.message}</p>
-      </div>
-    );
+    return ErrorMessage(error.message);
   }
+
   return (
     <div className="bg-white dark:bg-slate-900">
       <DataGrid
